LPD-47925 Move methods to frontend-js-components-web
