+++ modules/apps/frontend-js/frontend-js-components-web/package.json
@@ -22,7 +22,9 @@
 		"@liferay/frontend-js-state-web": "*",
 		"classnames": "2.3.1",
 		"frontend-js-web": "*",
+		"prop-types": "15.7.2",
 		"react": "18.2.0",
+		"react-dom": "18.2.0",
 		"uuid": "9.0.0"
 	},
 	"main": "index.ts",
