+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/index.ts
@@ -28,6 +28,23 @@ export {
 
 export {default as ManagementToolbar} from './management_toolbar/ManagementToolbar';
 
+export {default as openAlertModal} from './modal/commands/openAlertModal';
+export {default as openCategorySelectionModal} from './modal/commands/openCategorySelectionModal';
+export {default as openConfirmModal} from './modal/commands/openConfirmModal';
+export {default as openModal} from './modal/commands/openModal';
+export {default as openPortletModal} from './modal/commands/openPortletModal';
+export {default as openPortletWindow} from './modal/commands/openPortletWindow';
+export {
+	default as openSelectionModal,
+	OpenSelectionModalSelectedItem,
+} from './modal/commands/openSelectionModal';
+export {default as openSimpleInputModal} from './modal/commands/openSimpleInputModal';
+export {default as openTagSelectionModal} from './modal/commands/openTagSelectionModal';
+
+export {ModalStatus} from './modal/components/Modal';
+
+export {default as openToast} from './toast/openToast';
+
 export {Locale} from './translation_manager/TranslationAdminContent';
 export {default as TranslationAdminItem} from './translation_manager/TranslationAdminItem';
 export {default as TranslationAdminModal} from './translation_manager/TranslationAdminModal';
@@ -41,17 +58,3 @@ export {
 	selectedLanguageIdAtom,
 } from './translation_manager/state';
 export {default as Treeview} from './treeview/Treeview';
-
-export {
-	ModalStatus,
-	openAlertModal,
-	openCategorySelectionModal,
-	openConfirmModal,
-	openModal,
-	openPortletModal,
-	openPortletWindow,
-	openSelectionModal,
-	openSimpleInputModal,
-	openTagSelectionModal,
-	openToast,
-} from 'frontend-js-web';
