+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/modal/commands/openModal.ts
@@ -1,0 +1,30 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {render} from '@liferay/frontend-js-react-web';
+
+import Modal, {ModalProps} from '../components/Modal';
+
+export interface OpenModalProps extends ModalProps {}
+
+export default function openModal(props: OpenModalProps) {
+	if (
+		props &&
+		props.url &&
+		props.bodyHTML &&
+		process.env.NODE_ENV === 'development'
+	) {
+		console.warn(
+			'url and bodyHTML props are both set. bodyHTML will be ignored. Please use one or another.'
+		);
+	}
+
+	// Mount in detached node; Clay will take care of appending to `document.body`.
+	// See: https://github.com/liferay/clay/blob/master/packages/clay-shared/src/Portal.tsx
+
+	// @ts-ignore
+
+	render(Modal, props, document.createElement('div'));
+}
