+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/modal/components/Modal.scss
