+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/toast/notification.scss
