+++ modules/apps/frontend-js/frontend-js-components-web/test/tsconfig.json
@@ -1,0 +1,25 @@
+{
+	"@generated": "6c6c1748075cecbcf99f50a9a8f28b8dcd4d16a1",
+	"@readonly": "** AUTO-GENERATED: DO NOT EDIT **",
+	"compilerOptions": {
+		"allowSyntheticDefaultImports": true,
+		"baseUrl": ".",
+		"checkJs": false,
+		"composite": true,
+		"jsx": "react",
+		"module": "ESNext",
+		"moduleResolution": "node",
+		"rootDir": "../",
+		"strict": true,
+		"target": "es2020",
+		"typeRoots": [
+			"../../../../node_modules/@types"
+		]
+	},
+	"include": [
+		"**/*.ts",
+		"**/*.tsx",
+		"../src/**/*.ts",
+		"../src/**/*.tsx"
+	]
+}
