+++ modules/apps/frontend-js/frontend-js-web/package.json
@@ -1,17 +1,8 @@
 {
 	"dependencies": {
-		"@clayui/alert": "3.129.1",
-		"@clayui/button": "3.128.0",
-		"@clayui/form": "3.128.0",
-		"@clayui/icon": "3.128.0",
-		"@clayui/loading-indicator": "3.128.0",
-		"@clayui/modal": "3.128.0",
 		"@liferay/accessibility-settings-state-web": "*",
-		"@liferay/frontend-js-react-web": "*",
 		"@liferay/frontend-js-state-web": "*",
 		"@liferay/js-api": "0.6.0",
-		"classnames": "2.3.1",
-		"prop-types": "15.7.2",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
 		"uuid": "9.0.0"
