+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/global.es.js
@@ -99,8 +99,6 @@ import toggleRadio from './util/toggle_radio';
 import toggleSelectBox from './util/toggle_select_box';
 import zIndex from './zIndex';
 
-const PATH_CONTEXT = Liferay.ThemeDisplay.getPathContext();
-
 Liferay = window.Liferay || {};
 
 /**
@@ -158,18 +156,6 @@ Liferay.Portlet = Liferay.Portlet || {};
 
 Liferay.Portlet.minimize = minimizePortlet;
 
-Liferay.Portlet.openModal = (...args) => {
-	import(PATH_CONTEXT + '/o/frontend-js-web/__liferay__/index.js').then(
-		({openPortletModal}) => openPortletModal(...args)
-	);
-};
-
-Liferay.Portlet.openWindow = (...args) => {
-	import(PATH_CONTEXT + '/o/frontend-js-web/__liferay__/index.js').then(
-		({openPortletWindow}) => openPortletWindow(...args)
-	);
-};
-
 Liferay.SideNavigation = SideNavigation;
 
 Liferay.Util = Liferay.Util || {};
@@ -180,18 +166,6 @@ Liferay.Util.MAP_HTML_CHARS_ESCAPED = MAP_HTML_CHARS_ESCAPED;
  * @deprecated As of Athanasius (7.3.x), replaced by `import {addParams} from 'frontend-js-web'`
  */
 Liferay.Util.addParams = addParams;
-
-Liferay.Util.openAlertModal = (...args) => {
-	import(PATH_CONTEXT + '/o/frontend-js-web/__liferay__/index.js').then(
-		({openAlertModal}) => openAlertModal(...args)
-	);
-};
-
-Liferay.Util.openSimpleInputModal = (...args) => {
-	import(PATH_CONTEXT + '/o/frontend-js-web/__liferay__/index.js').then(
-		({openSimpleInputModal}) => openSimpleInputModal(...args)
-	);
-};
 
 /**
  * Utils added to global namespace to be consumed by portal-web
@@ -312,30 +286,6 @@ Liferay.Util.toCharCode = toCharCode;
  * @deprecated As of Athanasius (7.3.x), replaced by `import {toggleDisabled} from 'frontend-js-web'`
  */
 Liferay.Util.toggleDisabled = toggleDisabled;
-
-Liferay.Util.openConfirmModal = (...args) => {
-	import(PATH_CONTEXT + '/o/frontend-js-web/__liferay__/index.js').then(
-		({openConfirmModal}) => openConfirmModal(...args)
-	);
-};
-
-Liferay.Util.openModal = (...args) => {
-	import(PATH_CONTEXT + '/o/frontend-js-web/__liferay__/index.js').then(
-		({openModal}) => openModal(...args)
-	);
-};
-
-Liferay.Util.openSelectionModal = (...args) => {
-	import(PATH_CONTEXT + '/o/frontend-js-web/__liferay__/index.js').then(
-		({openSelectionModal}) => openSelectionModal(...args)
-	);
-};
-
-Liferay.Util.openToast = (...args) => {
-	import(PATH_CONTEXT + '/o/frontend-js-web/__liferay__/index.js').then(
-		({openToast}) => openToast(...args)
-	);
-};
 
 Liferay.Util.openWindow = openWindow;
 Liferay.Util.removeEntitySelection = removeEntitySelection;
