+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/liferay.d.ts
@@ -16,6 +16,10 @@ declare module Liferay {
 		): Promise<any>;
 	}
 
+	namespace CustomDialogs {
+		export const enabled: boolean;
+	}
+
 	namespace DOMTaskRunner {
 		export function addTask(task: object): void;
 
@@ -536,71 +540,6 @@ declare module Liferay {
 			object: Object
 		): URLSearchParams;
 
-		export function openModal(props: Object): void;
-
-		export function openPortletModal(
-			containerProps: Object,
-			footerCssClass: string,
-			headerCssClass: string,
-			iframeBodyCssClass: string,
-			onClose: () => void,
-			portletSelector: string,
-			subTitle: string,
-			title: string,
-			url: string
-		): void;
-
-		export function openSelectionModal(
-			buttonAddLabel: string,
-			buttonCancelLabel: string,
-			containerProps: Object,
-			customSelectEvent: boolean,
-			height: string,
-			id: string,
-			iframeBodyCssClass: string,
-			multiple: boolean,
-			onClose: () => void,
-			onSelect: () => void,
-			selectEventName: string,
-			selectedData: any,
-			size: 'full-screen' | 'lg' | 'md' | 'sm',
-			title: string,
-			url: string,
-			zIndex: number
-		): void;
-
-		/**
-		 * Function that implements the Toast pattern, which allows to present feedback
-		 * to user actions as a toast message in the lower left corner of the page
-		 */
-		export function openToast({
-			autoClose,
-			container,
-			containerId,
-			message,
-			onClick,
-			onClose,
-			renderData,
-			title,
-			toastProps,
-			type,
-			variant,
-		}: {
-			autoClose?: number | boolean;
-			container?: HTMLElement;
-			containerId?: string;
-			message?: string;
-			onClick?: () => void;
-			onClose?: () => void;
-			renderData?: {portletId: string};
-			title?: string;
-			toastProps?: Object;
-			type?: string;
-			variant?: string;
-		}): void;
-
-		export function openWindow(config: object, callback?: Function): void;
-
 		/**
 		 * Submits the form, with optional setting of form elements.
 		 */
@@ -710,8 +649,6 @@ declare module Liferay {
 		filterFn?: (component: any, componentConfigs: any) => boolean
 	): void;
 
-	export function detach(event: string, callback?: () => void): void;
-
 	/**
 	 * Clears the component promises map to make sure pending promises don't get
 	 * accidentally resolved at a later stage if a component with the same ID
@@ -733,13 +670,12 @@ declare module Liferay {
 
 	export function namespace(object: Object, path: string): Object;
 
-	export function on(events: string | string[], callback?: () => void): void;
-
 	export function SideNavigation(
 		toggler: HTMLElement,
 		options?: Object
 	): void;
 }
+
 interface ThemeDisplay {
 	isSignedIn(): boolean;
 	isStatePopUp(): boolean;
