+++ portal-web/docroot/html/common/themes/top_js.jspf
@@ -7,6 +7,21 @@
 
 <aui:script senna="temporary" type="text/javascript">
 	// <![CDATA[
+		function buildESMStub(contextPath, symbol) {
+			return (
+				(...args) => {
+					import(
+						Liferay.ThemeDisplay.getPathContext() +
+							'/o/' +
+							contextPath +
+							'/__liferay__/index.js'
+					).then(
+						(exports) => exports[symbol](...args)
+					);
+				}
+			);
+		}
+
 		var Liferay = Liferay || {};
 
 		Liferay.Browser = {
@@ -163,6 +178,13 @@
 				...acc, [key]: featureFlags[key] === 'true' || featureFlags[key] === true
 			}), {}
 		);
+
+		if (!Liferay.Portlet) {
+			Liferay.Portlet = {};
+		}
+
+		Liferay.Portlet.openModal = buildESMStub('frontend-js-components-web', 'openPortletModal');
+		Liferay.Portlet.openWindow = buildESMStub('frontend-js-components-web', 'openPortletWindow');
 
 		Liferay.PortletKeys = {
 			DOCUMENT_LIBRARY: '<%= PortletKeys.DOCUMENT_LIBRARY %>',
@@ -356,6 +378,17 @@
 				return <%= themeDisplay.isStatePopUp() %>;
 			}
 		};
+
+		if (!Liferay.Util) {
+			Liferay.Util = {};
+		}
+
+		Liferay.Util.openAlertModal = buildESMStub('frontend-js-components-web', 'openAlertModal');
+		Liferay.Util.openConfirmModal = buildESMStub('frontend-js-components-web', 'openConfirmModal');
+		Liferay.Util.openModal = buildESMStub('frontend-js-components-web', 'openModal');
+		Liferay.Util.openSelectionModal = buildESMStub('frontend-js-components-web', 'openSelectionModal');
+		Liferay.Util.openSimpleInputModal = buildESMStub('frontend-js-components-web', 'openSimpleInputModal');
+		Liferay.Util.openToast = buildESMStub('frontend-js-components-web', 'openToast');
 
 		var themeDisplay = Liferay.ThemeDisplay;
